+++ Scripts/Names/InputButtonName.cs
@@ -1,0 +1,26 @@
+namespace DSC.Input
+{
+    public struct InputButtonName
+    {
+        public const string axis = "Axis";
+        public const string axis2 = "Axis2";
+        public const string dPadUp = "DPadUp";
+        public const string dPadDown = "DPadDown";
+        public const string dPadLeft = "DPadLeft";
+        public const string dPadRight = "DPadRight";
+        public const string north = "North";
+        public const string south = "South";
+        public const string west = "West";
+        public const string east = "East";
+        public const string l1 = "L1";
+        public const string l2 = "L2";
+        public const string l3 = "L3";
+        public const string r1 = "R1";
+        public const string r2 = "R2";
+        public const string r3 = "R3";
+        public const string start = "Start";
+        public const string select = "Select";
+        public const string confirm = "Confirm";
+        public const string cancel = "Cancel";
+    }
+}